+++ src/domain/entity/place_plant.entity.ts
@@ -2,7 +2,7 @@ import { Column, Entity, PrimaryGeneratedColumn } from 'typeorm';
 
 @Entity({ name: 'local_planta' })
 export class PlacePlant {
-  @PrimaryGeneratedColumn({ name: 'local_planta_id', type: 'bigint' })
+  @PrimaryGeneratedColumn({ name: 'local_planta_id', type: 'int' })
   id: number;
 
   @Column({ name: 'planta' })
