+++ src/domain/entity/unity.entity.ts
@@ -2,7 +2,7 @@ import { Column, Entity, PrimaryGeneratedColumn } from 'typeorm';
 
 @Entity({ name: 'unidade_cliente' })
 export class Unity {
-  @PrimaryGeneratedColumn({ name: 'unidade_cliente_id', type: 'bigint' })
+  @PrimaryGeneratedColumn({ name: 'unidade_cliente_id', type: 'int' })
   id: number;
 
   @Column({ name: 'cnpj' })
